Make everything fully functional

+++ server/src/notes/notes.service.ts
@@ -1,4 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  HttpException,
+  HttpCode,
+  HttpStatus,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -15,17 +20,26 @@ export class NotesService {
     return this.notesRepository.find();
   }
 
-  public getOneNote(id: number) {
-    return this.notesRepository.findOne({ id });
-  }
+  public async getOneNote(id: number) {
+    const note = await this.notesRepository.findOne({ id });
 
-  public async createNote(note: Partial<Note>) {
-    return this.notesRepository.save(
-      await this.notesRepository.create(note)
+    if (note) {
+      return note;
+    }
+
+    throw new HttpException(
+      `Note with the ID of ${id} was not found`,
+      HttpStatus.NOT_FOUND,
     );
   }
 
+  public async createNote(note: Partial<Note>) {
+    return this.notesRepository.save(this.notesRepository.create(note));
+  }
+
   public async updateNote(id: number, note: Partial<Note>) {
+    await this.getOneNote(id);
+
     const { createdAt, updatedAt, ...restOfNote } = note;
     await this.notesRepository.update(id, restOfNote);
 
@@ -33,10 +47,12 @@ export class NotesService {
   }
 
   public async deleteNote(id: number) {
-    await this.notesRepository.delete(id);
+    if (await this.getOneNote(id)) {
+      await this.notesRepository.delete(id);
 
-    return {
-      deleted: true,
-    };
+      return {
+        deleted: true,
+      };
+    }
   }
 }
